requirements.txt
pygame==2.5.2
requests==2.31.0
PyGithub==2.1.1
RPi.GPIO==0.7.1 